+++ src/App/Models/Database.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Controllers;
+namespace App\Models;
 
 use PDO;
