+++ src/App/Models/Model.php
@@ -2,7 +2,7 @@
 
 namespace App\Models;
 
-use App\Controllers\Database;
+use App\Models\Database;
 
 abstract class Model
 {
@@ -16,25 +16,13 @@ abstract class Model
     }
 
     /**
-     * find
-     */
-    public function find(int $id) : array
-    {
-
-        $query = $this->pdo->prepare("SELECT * FROM {$this->table} WHERE 'id' = :id");
-        $query->execute(['id' => $id]);
-        $item = $query->fetch();
-
-        return $item;
-    }
-
-    /**
      * delete
+     *
+     *    public function delete(int $id) :void
+     *    {
+     *
+     *        $query = $this->pdo->prepare("DELETE FROM {$this->table} WHERE id = :id");
+     *        $query->execute(['id' => $id]);
+     *    }
      */
-    public function delete(int $id) :void
-    {
-
-        $query = $this->pdo->prepare("DELETE FROM {$this->table} WHERE id = :id");
-        $query->execute(['id' => $id]);
-    }
 }
